+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Chg #207: Replace two `RouteCollectorInterface` methods `addRoute()` and `addGroup()` to single `addRoute()` (@vjik)
 - Chg #222: Make `Route`, `Group` and `MatchingResult` dispatcher-independent (@rustamwin, @vjik)
 - Enh #229: Add URL arguments' psalm type in `UrlGeneratorInterface` (@vjik)
+## 3.1.1 under development
+
+- no changes in this release.
 
 ## 3.1.0 February 20, 2024